Enable auto-merge for Flathub bot PRs
VS Code Insiders is very much at-your-own-risk software; it is in a
separate beta-only stream, parallel-installable with the regular edition
of Code, and is not necessarily suitable for day-to-day use.

Given that, and the volume of updates, enable auto-merging of Flathub
bot PRs, so people can live even closer to the bleeding edge.